added SD card logging
got the SD card logging working but the concat of all the data into the
string isnt working, only getting a bunch of commas instead of the data
separated by commas, still gotta figure it out. Something to do with
String() not making those values into a string? maybe dont use it?